refactor:list.js 파일 분리 및 html 구조 수정
