+++ pages/list/list.html
@@ -1,37 +1,20 @@
-<!doctype html>
-<html lang="ko-kr">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=1500, user-scalable=no" />
-    <title>커뮤니티 - CoDive</title>
-    <link rel="stylesheet" href="../../reset.css" />
-    <link rel="stylesheet" href="./list.css" />
-    <script type="module" src="./list.js"></script>
-  </head>
-  <body>
-    <!-- 네브바 -->
+<!-- 헤로우 -->
 
-    <!-- 헤로우 -->
-    <div class="hero">여러분의 경험이 누군가의 성장의 씨앗이 됩니다 ✍️</div>
+<!-- 본문 컨테이너 -->
+<div class="container">
+  <div class="hero">여러분의 경험이 누군가의 성장의 씨앗이 됩니다 ✍️</div>
+  <div class="button-group">
+    <button id="free-button" class="button-active">free</button>
+    <button id="group-button">group</button>
+  </div>
 
-    <!-- 본문 컨테이너 -->
-    <div class="container">
-      <div class="button-group">
-        <button id="free-button" class="button-active">write</button>
-        <button id="group-button">preview</button>
-      </div>
+  <!-- 글 목록 -->
+  <ul class="list">
+    <!-- 예시 목록 10개 -->
+  </ul>
 
-      <!-- 글 목록 -->
-      <ul class="list">
-        <!-- 예시 목록 10개 -->
-      </ul>
-
-      <!-- 페이지네이션 -->
-      <div class="pagination">
-        <span>1</span>
-      </div>
-    </div>
-
-    <!-- 푸터 -->
-  </body>
-</html>
+  <!-- 페이지네이션 -->
+  <div class="pagination">
+    <span>1</span>
+  </div>
+</div>
